Move is_subsequence to utils.py Update comments to be send marker string generic Reformat extra args to be more ledgible
