+++ tests/integrationv2/utils.py
@@ -7,6 +7,38 @@ def to_bytes(val):
     return bytes(str(val).encode('utf-8'))
 
 
+def is_subsequence(s, t):
+    """predicate: is s a subsequence of t?
+
+    Expected Usage
+    --------------
+    Was the expected data printed to stdout (even if non-contiguously)?
+    
+        assert(is_subsequence(data, stdout))
+
+    Parameters
+    ----------
+    s: bytes
+        subsequence to look for
+    t: str, bytes
+        sequence to look in
+
+    Returns
+    -------
+    bool
+        True if s is a subsequence, False if not 
+    """
+    s = str(s)[2:-1] # Remove the b' and ' from the converted bytes
+    t = str(t)
+    s_len, t_len = len(s), len(t)
+    s_index, t_index = 0, 0
+    while (s_index < s_len and t_index < t_len):
+        if (s[s_index] == t[t_index]):
+            s_index += 1
+        t_index += 1
+    return s_index == s_len
+
+
 def get_expected_s2n_version(protocol, provider):
     """
     s2nd and s2nc print a number for the negotiated TLS version.
